+++ src/AwesomeApp.UI/src/backend/settings/index.ts
@@ -1,5 +1,5 @@
 import { ApiSettings } from './ApiSettings'
-import { AppSettings } from '@/shared/models/AppSettings'
+import { AppSettings } from '@/shared/types/AppSettings'
 import apiLocalSettigns from './api.local.settings.json'
 import apiProdSettigns from './api.local.settings.json'
 import appLocalSettigns from './api.local.settings.json'
