+++ src/AwesomeApp.UI/src/pages/api/account/authenticate.ts
@@ -1,5 +1,5 @@
 import { withAuthentication, withEndpoints, withErrorHandling } from "@/backend/libs";
-import { HttpMethod } from "@/shared/models/HttpMethod";
+import { HttpMethod } from "@/shared/types/HttpMethod";
 import { postAuthenticate } from "@/backend/handlers/AccountHandlers";
 
 export default
