+++ src/AwesomeApp.UI/src/shared/dtos/index.ts
@@ -1,4 +1,4 @@
 export * from './AccountDto'
+export * from './AuthenticateAccountDto'
 export * from './AuthenticationResultDto'
 export * from './CreateAccountDto'
-export * from './AuthenticateAccountDto'
