+++ src/AwesomeApp.UI/src/shared/types/AppSettings.d.ts
