+++ src/AwesomeApp.UI/src/shared/types/CustomerType.ts
