+++ src/AwesomeApp.UI/src/shared/types/HttpMethod.ts
