+++ src/AwesomeApp.UI/src/shared/types/index.ts
